Use shared doc comment manager in feather fixes
